pyproject.toml
[project]
name = "ai-news-blog-automation"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "feedparser>=6.0.12",
    "openai>=2.21.0",
    "pillow>=12.1.1",
    "praw>=7.8.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "textstat>=0.7.13",
    "tweepy>=4.16.0",
]
